pyproject.toml
[project]
name = "ecommerce-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "crewai>=0.193.2",
    "google-genai>=1.69.0",
    "langchain>=1.2.13",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
]
